backend/pyproject.toml
[project]
name = "my-library-backend"
version = "0.1.0"
description = "マンガ管理アプリケーションのバックエンドAPI"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "D",   # pydocstyle (docstrings)
]
ignore = [
    "E501",  # line too long (handled by black)
    "D100",  # public module docstring (MVPでは必須にしない)
    "D104",  # public package docstring (MVPでは必須にしない)
    "D107",  # __init__ docstring (MVPでは必須にしない)
]

[tool.ruff.lint.isort]
known-first-party = ["my-library-backend"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.black]
line-length = 100
target-version = ["py39"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = "(\\.venv|\\.uv|\\.mypy_cache|node_modules)"
